requirements.txt
# Core dependencies
apache-airflow==2.8.0
pyspark==3.5.0
requests==2.31.0
pandas==2.1.4

# Data processing
pyarrow==14.0.1
fastparquet==2023.10.1

# Cloud storage
google-cloud-storage==2.10.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
requests-mock==1.11.0

# Code quality
black==23.12.1
flake8==7.0.0
mypy==1.7.1
isort==5.13.2

# Typing
types-requests==2.31.0.10
